Authentication

+++ PycharmProjects/python_learning/Fast_API/TodoApp/routers/todos.py
@@ -2,10 +2,10 @@ from typing import Annotated
 from sqlalchemy.orm import Session
 from fastapi import APIRouter, Depends, HTTPException, Path
 from starlette import status
-
 from models import Todos
 from database import SessionLocal
 from pydantic import BaseModel,Field
+from .auth import get_current_user
 
 
 router = APIRouter()
@@ -23,6 +23,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str=Field(min_length=3)
@@ -31,29 +32,46 @@ class TodoRequest(BaseModel):
     complete : bool
 
 
-@router.get("/")
-async def read_all(db: db_dependency ):
-    return db.query(Todos).all()
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(user: user_dependency, db: db_dependency ):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_id(db: db_dependency, todo_id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_id(user: user_dependency, db: db_dependency, todo_id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo is not found")
 
+
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(todo_request: TodoRequest,
+                      user: user_dependency,
+                      db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
+
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
-                      todo_request: TodoRequest,
+async def update_todo( user: user_dependency,
+                      db: db_dependency,
+                       todo_request: TodoRequest,
                       todo_id: int=Path(gt=0),
                       ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="No Id found in data")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="No Id found in data")
 
@@ -65,14 +83,21 @@ async def update_todo(db: db_dependency,
     db.add(todo_model)
     db.commit()
 
+
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id : int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency,
+                      db: db_dependency,
+                      todo_id : int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=404, detail="Id not found to delete")
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id')).first())
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Id not found to delete")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
